apps/sof15/requirements.txt
Django==1.7.1
django-auth-ldap==1.2.3
django-localflavor==1.0
django-crispy-forms==1.4.0
django-mptt==0.6.1
liu==1.0.0
pkginfo==1.1
python-ldap==2.4.18
raven==5.1.1
requests==2.5.0
wsgiref==0.1.2
